+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to Pepper Terminal will be documented here.
 
 ## [Unreleased]
-- TBD
+- “Rotate memory logs daily (one JSONL file per date).”
 
 ## [0.1.0] - 2025-12-12
 ### Added